修正分数定义

+++ src/com/xuexin/wangshen/util/calmarks/MarksCaculator1IsFilled.java
@@ -26,7 +26,7 @@ public class MarksCaculator1IsFilled extends MarksCaculator {
 		}
 		
 		//其他情况得分
-		return (double)objJudge.getIntValue("score");
+		return (double)objJudge.getFloatValue("score");
 	}
 
 }
