+++ src/com/xuexin/wangshen/util/calmarks/MarksCaculator3IsIDCard.java
@@ -37,7 +37,7 @@ public class MarksCaculator3IsIDCard extends MarksCaculator {
         }
 		
 		//其他情况得分
-		return (double)objJudge.getIntValue("score");
+		return (double)objJudge.getFloatValue("score");
 	}
 
 }
